refactor: reorganize HTTP package and improve API usability
- Rename DoHTTPRequestUnifiedWithRetry to DoHTTPRequestWithRetry
- Remove DoHTTPRequestUnified (make retry the default behavior)
- Organize code into focused modules:
  - client.go: HTTP client configuration
  - latency.go: latency tracking and metrics
  - processing.go: cookie/header processing
  - common.go: core retry logic
- Update all callers to use retry by default
- Clean up unused functions and imports
