+++ umh-core/pkg/communicator/api/v2/http/requester_test.go
@@ -43,7 +43,7 @@ var _ = Describe("Requester", func() {
 	doHTTPRequestWithRetry := func(ctx context.Context, url string, header map[string]string, cookies *map[string]string, insecureTLS bool, logger *zap.SugaredLogger) (*netHTTP.Response, error) {
 		var lastErr error
 		for i := range 10 {
-			response, err := http.DoHTTPRequestUnified[any](ctx, netHTTP.MethodGet, url, nil, header, cookies, insecureTLS, true, logger)
+			response, err := http.DoHTTPRequestWithRetry[any](ctx, netHTTP.MethodGet, url, nil, header, cookies, insecureTLS, true, logger)
 			if err == nil {
 				return response, nil
 			}
